+++ src/main/java/com/spikeify/commands/SingleObjectDeleter.java
@@ -42,7 +42,7 @@ public class SingleObjectDeleter {
 		if (meta.userKeyString != null) {
 			return new Key(meta.namespace, meta.setName, meta.userKeyString);
 		} else {
-			return new Key(meta.namespace, meta.setName, meta.userKeyLong.longValue());
+			return new Key(meta.namespace, meta.setName, meta.userKeyLong);
 		}
 	}
