+++ src/main/java/com/spikeify/commands/SingleObjectUpdater.java
@@ -84,7 +84,7 @@ public class SingleObjectUpdater<T> {
 		if (meta.userKeyString != null) {
 			return new Key(meta.namespace, meta.setName, meta.userKeyString);
 		} else {
-			return new Key(meta.namespace, meta.setName, meta.userKeyLong.longValue());
+			return new Key(meta.namespace, meta.setName, meta.userKeyLong);
 		}
 	}
